// src/enums/endpoints.enum.ts
export enum EndpointsEnum {
    AUTH = 'auth',
    USER = 'users',
    ROLE = 'roles',
    PERMISSION = 'permissions',
    PRICE_CATALOG = 'price-catalog',
    REVENUE = 'revenues',
    EXPENSE = 'expenses',
    PATIENT = 'patients',
    VISIT = 'visits',
    SETTINGS = 'settings',
    FINANCE = 'finance',
    MEDICAL_DOCUMENTS = 'medical-documents',
}